Fix vis_writer for latest ChunkStore and L0 stream
The ChunkStore put primitive changed to put_chunk.

The `dump_index` SPEAD item in the latest sdp-l0 stream is of type
np.uint64. Somewhat unexpectedly, adding an integer to it promotes
the result to a float (check type(np.uint64(0) + 1) for yourself).
Also see NumPy GitHub issues #3118, #5745, #7126 and several more.
The solution is to cast dump_index to an int - it is not expected
to get mega-large.