Run dask on large enough graphs, and profile
Instead of running the dask scheduler on a dump at a time, rather
collect the tasks in a pending graph until we have enough (typically
around 200) and only then write it out to the chunk store. This
fixes support for fast dump rates (i.e. > 0.25 Hz) coupled with
low data rates, which result in only a handful of chunks per dump.
Also add profiling to check data rates, which will be useful once
we swop out the hand-crafted dask graph for a standard dask Array
to be used with the pending `ChunkStore.put_dask_array` method,
just to check that we are not impacting performance.

Allocate SPEAD ring buffers and memory pools based on the expected
number of dumps per write cycle.