flag_writer: handle do_capture internally
It was previously left to the script to run do_capture and wait for it.
It's now run as a task internally. That makes the server more of a
self-contained black box, and potentially fixes some shutdown ordering
bugs where the server would close down its connections before it was
done clearing the heaps in its queue.