Redesign dashboard to not use periodic sampling
Instead, react to changes on the sensors. Use a separate
ColumnDataSource per sensor since their timestamps won't align. This
allows arbitrarily high resolution, although that can also be a bad
thing as it can store too many samples for the browser to cope with.

Also change the graphs to be step graphs since that reflects the sensor
values at intermediate times.