Add sensors to track write queue depth
This is separate from the active-chunks sensor, which indicates the
number of chunks that are actually inside write() calls. It includes
data that has been received but hasn't yet been taken by a worker.

It's specified in bytes rather than chunks because it is not clearly
aligned to either input or output chunks: it increments in input chunks
but decrements in output chunks. Once we start doing time accumulation,
neither is a subset of the other.