Rework buffering in datawriter
Instead of having a big spead2 ringbuffer, buffer data in the queue of
the executor that handles the writes. Moving the buffer down the
pipeline allows the earlier stages to keep making progress for a while
even when the disk system stalls out, so that when it recovers we
already have a queue of data ready to write. It's also easier to monitor
the backpressure there rather than trying to reach into C++ internals.

This required a more sophisticated accounting for the queue, because
it contains chunks of different sizes. Thus, a bound on items in the
queue doesn't directly bound memory usage, since it depends on the mix
of items in the queue. The semaphore that used to bound it is thus
replaced by a hand-written class (QueueSpace) that generalises a
semaphore to allow acquiring and releasing byte counts instead of
unitary items.

The default buffer size is quite a bit larger than before, in an attempt
to buffer over disk stalls due to SR-1505.

Relates to SR-1459.