Limit --obj-max-dumps to not be bigger than --buffer-dumps
This addresses two issues:
1. In extreme cases, an output dump could be larger than the write
buffer, which would then deadlock because output dumps could never be
added to the buffer.
2. Even if the above does not occur, there are likely to be stalls and
dropped packets, because as soon as the first new data is seen for the
next output dump, the current dump has to be fed into the buffer and
that will need to wait for some writes to complete before it can all be
fed in. That will then block processing of new heaps. The network
buffers can absorb some of that, but the design is to put most of the
buffer space into the write buffers and keep network buffers just big
enough to cover jitter.