Ensure that queue counter arithmetic balances
Missing chunks were not being count towards increasing the queue size,
but were being counted when subtracting (as long as some other data in
the same output chunk was received). Fix it by keeping track in each
_Item of the input chunks that were received, so that the "payload"
quantity in each output chunk can be computed.

Also fixed an issue where the rechunker tests had a base class marked
`@nottest`, and that propagated into the derived classes that were supposed
to be tested. The `@nottest` decorator is removed and the base class
made private (`_`-prefixed) instead.

Closes SPR1-330.